+++ build.rs
@@ -1,5 +1,0 @@
-extern crate gcc;
-
-fn main () {
-    println!("cargo:rustc-flags=-l dylib=stdc++");
-}