+++ rpi-led-matrix/src/options.rs
@@ -1,42 +1,17 @@
-use libc::{c_char, c_int};
+use libc::c_int;
 use std::ffi::CString;
 
+use crate::ffi;
+
+/// Options related to the LED matrix, like rows/cols/HW mapping
+#[derive(Debug)]
+pub struct LedMatrixOptions(pub(crate) ffi::CLedMatrixOptions);
+
+/// Options related to how the runtime operates, like GPIO slowdown or daemon/sudo privileges
+#[derive(Debug)]
+pub struct LedRuntimeOptions(pub(crate) ffi::CLedRuntimeOptions);
+
 type LedMatrixOptionsResult = Result<(), &'static str>;
-
-/// The Rust representation of LedMatrixOptions, which contains parameters to specify your hardware setup.
-#[derive(Debug)]
-#[repr(C)]
-pub struct LedMatrixOptions {
-    pub(crate) hardware_mapping: *mut c_char,
-    pub(crate) rows: c_int,
-    pub(crate) cols: c_int,
-    pub(crate) chain_length: c_int,
-    pub(crate) parallel: c_int,
-    pub(crate) pwm_bits: c_int,
-    pub(crate) pwm_lsb_nanoseconds: c_int,
-    pub(crate) pwm_dither_bits: c_int,
-    pub(crate) brightness: c_int,
-    pub(crate) scan_mode: c_int,
-    pub(crate) row_address_type: c_int,
-    pub(crate) multiplexing: c_int,
-    pub(crate) led_rgb_sequence: *mut c_char,
-    pub(crate) pixel_mapper_config: *mut c_char,
-    pub(crate) panel_type: *mut c_char,
-    pub(crate) disable_hardware_pulsing: c_char,
-    pub(crate) show_refresh_rate: c_char,
-    pub(crate) inverse_colors: c_char,
-    pub(crate) limit_refresh_rate_hz: c_int,
-}
-
-/// The Rust representation of LedRuntimeOptions, which contains parameters to specify how the library behaves at runtime.
-#[derive(Debug)]
-#[repr(C)]
-pub struct LedRuntimeOptions {
-    pub(crate) gpio_slowdown: c_int,
-    pub(crate) daemon: c_int,
-    pub(crate) drop_privileges: c_int,
-    pub(crate) do_gpio_init: bool,
-}
 
 impl LedMatrixOptions {
     /// Creates a new `LedMatrixOptions` struct with the default parameters.
@@ -47,8 +22,8 @@ impl LedMatrixOptions {
     /// options.set_hardware_mapping("adafruit-hat-pwm");
     /// let matrix = LedMatrix::new(Some(options), None).unwrap();
     /// ```
-    pub fn new() -> LedMatrixOptions {
-        LedMatrixOptions {
+    pub fn new() -> Self {
+        Self(ffi::CLedMatrixOptions {
             hardware_mapping: CString::new("regular").unwrap().into_raw(),
             rows: 32,
             cols: 32,
@@ -68,35 +43,35 @@ impl LedMatrixOptions {
             show_refresh_rate: 1,
             inverse_colors: 0,
             limit_refresh_rate_hz: 0,
-        }
+        })
     }
 
     /// Sets the type of GPIO mapping used (e.g., "adafruit-hat-pwm").
     pub fn set_hardware_mapping(&mut self, mapping: &str) {
         unsafe {
-            let _ = CString::from_raw(self.hardware_mapping);
-            self.hardware_mapping = CString::new(mapping).unwrap().into_raw();
+            let _ = CString::from_raw(self.0.hardware_mapping);
+            self.0.hardware_mapping = CString::new(mapping).unwrap().into_raw();
         }
     }
 
     /// Sets the number of rows on the panels being used. Typically 8, 16, 32 or 64.
     pub fn set_rows(&mut self, rows: u32) {
-        self.rows = rows as c_int;
+        self.0.rows = rows as c_int;
     }
 
     /// Sets the number of columns on the panels being used. Typically 32 or 64.
     pub fn set_cols(&mut self, cols: u32) {
-        self.cols = cols as c_int;
+        self.0.cols = cols as c_int;
     }
 
     /// Sets the number of panels daisy-chained together.
     pub fn set_chain_length(&mut self, chain_length: u32) {
-        self.chain_length = chain_length as c_int;
+        self.0.chain_length = chain_length as c_int;
     }
 
     /// Sets the number of parallel chains. Valid range: [1,3].
     pub fn set_parallel(&mut self, parallel: u32) {
-        self.parallel = parallel as c_int;
+        self.0.parallel = parallel as c_int;
     }
 
     /// Sets the number of PWM bits to use. Valid range: [1,11].
@@ -104,36 +79,36 @@ impl LedMatrixOptions {
         if pwm_bits > 11 {
             Err("Pwm bits can only have value between 0 and 11 inclusive")
         } else {
-            self.pwm_bits = pwm_bits as c_int;
+            self.0.pwm_bits = pwm_bits as c_int;
             Ok(())
         }
     }
 
     /// Sets the number of nanoseconds of delay for our LSB
     pub fn set_pwm_lsb_nanoseconds(&mut self, pwm_lsb_nanoseconds: u32) {
-        self.pwm_lsb_nanoseconds = pwm_lsb_nanoseconds as c_int;
+        self.0.pwm_lsb_nanoseconds = pwm_lsb_nanoseconds as c_int;
     }
 
     /// Sets the pannel brightness in percent.
     pub fn set_brightness(&mut self, brightness: u8) -> LedMatrixOptionsResult {
-        if brightness > 100 || brightness < 1 {
-            Err("Brigthness can only have value between 1 and 100 inclusive")
+        if !(1..=100).contains(&brightness) {
+            Err("Brightness can only have value between 1 and 100 inclusive")
         } else {
-            self.brightness = brightness as c_int;
+            self.0.brightness = brightness as c_int;
             Ok(())
         }
     }
 
     /// Sets the scan mode. 0: progressive, 1: interlaced.
     pub fn set_scan_mode(&mut self, scan_mode: u32) {
-        self.scan_mode = scan_mode as c_int;
+        self.0.scan_mode = scan_mode as c_int;
     }
 
     /// Sets the ordering of the LEDs on your panel.
     pub fn set_led_rgb_sequence(&mut self, sequence: &str) {
         unsafe {
-            let _ = CString::from_raw(self.led_rgb_sequence);
-            self.led_rgb_sequence = CString::new(sequence).unwrap().into_raw();
+            let _ = CString::from_raw(self.0.led_rgb_sequence);
+            self.0.led_rgb_sequence = CString::new(sequence).unwrap().into_raw();
         }
     }
 
@@ -147,35 +122,35 @@ impl LedMatrixOptions {
     /// * V-mapper
     pub fn set_pixel_mapper_config(&mut self, mapper: &str) {
         unsafe {
-            let _ = CString::from_raw(self.pixel_mapper_config);
-            self.pixel_mapper_config = CString::new(mapper).unwrap().into_raw();
+            let _ = CString::from_raw(self.0.pixel_mapper_config);
+            self.0.pixel_mapper_config = CString::new(mapper).unwrap().into_raw();
         }
     }
 
     /// Sets if hardware pin-pulse generation should be used.
     pub fn set_hardware_pulsing(&mut self, enable: bool) {
         if enable {
-            self.disable_hardware_pulsing = 0;
+            self.0.disable_hardware_pulsing = 0;
         } else {
-            self.disable_hardware_pulsing = 1;
+            self.0.disable_hardware_pulsing = 1;
         }
     }
 
     /// Configures if the current refresh rate should be printed by the C++ library.
     pub fn set_refresh_rate(&mut self, enable: bool) {
         if enable {
-            self.show_refresh_rate = 1;
+            self.0.show_refresh_rate = 1;
         } else {
-            self.show_refresh_rate = 0;
+            self.0.show_refresh_rate = 0;
         }
     }
 
     /// If set, invert the color displayed.
     pub fn set_inverse_colors(&mut self, enable: bool) {
         if enable {
-            self.inverse_colors = 1;
+            self.0.inverse_colors = 1;
         } else {
-            self.inverse_colors = 0;
+            self.0.inverse_colors = 0;
         }
     }
 
@@ -199,7 +174,7 @@ impl LedMatrixOptions {
     /// 15. P10CoremanMapper
     /// 16. P8Outdoor1R1G1
     pub fn set_multiplexing(&mut self, multiplexing: u32) {
-        self.multiplexing = multiplexing as c_int;
+        self.0.multiplexing = multiplexing as c_int;
     }
 
     /// Sets the type of row addressing to be used.
@@ -210,26 +185,26 @@ impl LedMatrixOptions {
     /// 3. ABC-addressed panels
     /// 4. ABC Shift + DE direct
     pub fn set_row_addr_type(&mut self, row_addr_type: u32) {
-        self.row_address_type = row_addr_type as c_int;
+        self.0.row_address_type = row_addr_type as c_int;
     }
 
     /// Limit refresh rate to this frequency in Hz. (0 = no limit)
     ///
     /// Useful to keep a constant refresh rate on loaded system.
     pub fn set_limit_refresh(&mut self, limit_refresh: u32) {
-        self.limit_refresh_rate_hz = limit_refresh as c_int;
+        self.0.limit_refresh_rate_hz = limit_refresh as c_int;
     }
 
     /// Configures how many bits to use for time-based dithering.
     pub fn set_pwm_dither_bits(&mut self, pwm_dither_bits: u32) {
-        self.pwm_dither_bits = pwm_dither_bits as c_int;
+        self.0.pwm_dither_bits = pwm_dither_bits as c_int;
     }
 
     /// Needed to initialize special panels. Supported: 'FM6126A', 'FM6127'
     pub fn set_panel_type(&mut self, panel_type: &str) {
         unsafe {
-            let _ = CString::from_raw(self.panel_type);
-            self.panel_type = CString::new(panel_type).unwrap().into_raw();
+            let _ = CString::from_raw(self.0.panel_type);
+            self.0.panel_type = CString::new(panel_type).unwrap().into_raw();
         }
     }
 }
@@ -243,9 +218,9 @@ impl Default for LedMatrixOptions {
 impl Drop for LedMatrixOptions {
     fn drop(&mut self) {
         unsafe {
-            let _ = CString::from_raw(self.hardware_mapping);
-            let _ = CString::from_raw(self.led_rgb_sequence);
-            let _ = CString::from_raw(self.panel_type);
+            let _ = CString::from_raw(self.0.hardware_mapping);
+            let _ = CString::from_raw(self.0.led_rgb_sequence);
+            let _ = CString::from_raw(self.0.panel_type);
         }
     }
 }
@@ -260,32 +235,32 @@ impl LedRuntimeOptions {
     /// let matrix = LedMatrix::new(None, Some(rt_options)).unwrap();
     /// ```
     pub fn new() -> Self {
-        Self {
+        Self(ffi::CLedRuntimeOptions {
             gpio_slowdown: 1,
             daemon: 0,
             drop_privileges: 1,
             do_gpio_init: true,
-        }
+        })
     }
 
     /// Sets the GPIO slowdown, in . Needed for faster Pis/slower panels
     pub fn set_gpio_slowdown(&mut self, gpio_slowdown: u32) {
-        self.gpio_slowdown = gpio_slowdown as i32;
+        self.0.gpio_slowdown = gpio_slowdown as i32;
     }
 
     /// If True, make the process run in the background as daemon.
     pub fn set_daemon(&mut self, daemon: bool) {
-        self.daemon = if daemon { 1 } else { 0 };
+        self.0.daemon = if daemon { 1 } else { 0 };
     }
 
     /// If True, drop privileges from 'root' after initializing the hardware.
     pub fn set_drop_privileges(&mut self, drop_privileges: bool) {
-        self.drop_privileges = if drop_privileges { 1 } else { 0 };
+        self.0.drop_privileges = if drop_privileges { 1 } else { 0 };
     }
 
     /// You almost definitely want this to be left as True. Use this if you know what you're doing.
     pub fn set_do_gpio_init(&mut self, do_gpio_init: bool) {
-        self.do_gpio_init = do_gpio_init;
+        self.0.do_gpio_init = do_gpio_init;
     }
 }